-update JR to 0.17 (as the one in CBA)
